requirements.txt
streamlit
pandas
PyPDF2
openpyxl
pinecone-client
google-genai
python-dotenv
